chore: reset kotlin to main for phase 0 rebuild
Hard-reset all Kotlin changes from this branch in preparation for the
phased rebuild against the v2 plan (share at coordinator level, defer
divergence to View). Branch commits remain on origin for reference.
JS work in descope-js is untouched.

See plans/look-at-https-github-com-descope-etc-iss-lovely-parasol.md
