+++ descopesdk/src/main/java/com/descope/android/DescopeFlowHook.kt
@@ -1,38 +1,22 @@
-package com.descope.android.bridge
+package com.descope.android
 
 import android.webkit.WebView
-import com.descope.android.bridge.DescopeBridgeHook.Event
-
-interface Coordinator {
-    /** The underlying [WebView] that hosts the flow or widget. */
-    val webView: WebView
-
-    /** Runs the given JavaScript code in the page, wrapped in an IIFE. */
-    fun runJavaScript(code: String)
-
-    /** Injects a `<style>` element into the page with the given CSS. */
-    fun addStyles(css: String)
-}
+import com.descope.android.DescopeFlowHook.Event
 
 /**
- * The [DescopeBridgeHook] class allows implementing hooks that customize how a Descope
- * bridge's webpage looks or behaves, usually by adding CSS, running JavaScript code, or
- * configuring the web view.
+ * The [DescopeFlowHook] class allows implementing hooks that customize how the flow
+ * webpage looks or behaves, usually by adding CSS, running JavaScript code, or configuring
+ * the web view.
  *
- * The same hook type is used by both [com.descope.android.flow.DescopeFlow] and
- * [com.descope.android.widget.DescopeUserProfileWidget] — both expose a `hooks` list and
- * both pass a [Coordinator] to your hook's [execute].
- *
- * You can use hooks by setting the bridge's `hooks` array. For example, these hooks will
- * override the flow or widget to have a transparent background and set margins on the body
- * element.
+ * You can use hooks by setting the [DescopeFlow.hooks] array. For example, these hooks will
+ * override the flow to have a transparent background and set margins on the body element.
  *
  *     flow.hooks = listOf(
  *         setTransparentBody(),
  *         addStyles(selector = "body", rules = listOf("margin: 16px")),
  *     )
  *
- * Alternatively, create custom hooks in a [DescopeBridgeHook] extension to have them all
+ * Alternatively, create custom hooks in a [DescopeFlowHook] extension to have them all
  * in one place:
  *
  *     fun showFlow() {
@@ -42,61 +26,61 @@ interface Coordinator {
  *     }
  *
  *     // elsewhere
- *
+ *     
  *     val setMaxWidth = addStyles(selector = ".login-container", rules = listOf("max-width: 250px"))
- *
- *     val removeFooter = runJavaScript(DescopeBridgeHook.Event.Ready, """
+ *     
+ *     val removeFooter = runJavaScript(DescopeFlowHook.Event.Ready, """
  *         const footer = document.querySelector('#footer')
  *         footer?.remove()
  *     """)
- *
+ *        
  *     val hideScrollBar = setUpWebView {
  *         it.isVerticalScrollBarEnabled = false
  *     }
  *
- * You can also implement your own hooks by subclassing [DescopeBridgeHook] and
- * overriding the [DescopeBridgeHook.execute] method.
- *
+ * You can also implement your own hooks by subclassing [DescopeFlowHook] and
+ * overriding the [DescopeFlowHook.execute] method.
+ * 
  * @param events A set of events for which the hook will be executed.
  */
-abstract class DescopeBridgeHook(val events: Set<Event>) {
+abstract class DescopeFlowHook(val events: Set<Event>) {
 
     /** The hook event determines when a hook is executed. */
     enum class Event {
         /**
-         * The hook is executed when the flow or widget is started.
+         * The hook is executed when the flow is started with `run()`.
          *
-         * - Note: The page is not loaded and the `document` element isn't available
+         * - Note: The flow is not loaded and the `document` element isn't available
          *     at this point, so this event is not appropriate for making changes to
-         *     the page itself.
+         *     the flow page itself.
          */
         Started,
 
         /** The hook is executed when the `document` element is available in the page. */
         Loaded,
 
-        /** The hook is executed when the page is fully loaded and ready to be displayed. */
+        /** The hook is executed when the flow page is fully loaded and ready to be displayed. */
         Ready,
     }
 
     /**
      * Override this abstract method to implement your hook.
      *
-     * This method is called when one of the events in the [events] set takes place. If
-     * the set has more than one member you can check the `event` parameter and take
-     * different actions depending on the specific event.
+     * This method is called by the [DescopeFlowCoordinator] when one of the events in
+     * the [events] set takes place. If the set has more than one member you can check
+     * the `event` parameter and take different actions depending on the specific event.
      *
      * @param event The event that took place.
-     * @param coordinator The [Coordinator] that's running the flow or widget.
+     * @param coordinator The [DescopeFlowCoordinator] that's running the flow.
      */
-    abstract fun execute(event: Event, coordinator: Coordinator)
+    abstract fun execute(event: Event, coordinator: DescopeFlowCoordinator)
 
     companion object {
         /**
          * The list of default hooks.
          *
          * These hooks are always executed, but you can override them by adding the
-         * counterpart hook to the bridge's `hooks` array.
+         * counterpart hook to the [DescopeFlow.hooks] array.
          */
         val defaults = listOf(
             disableZoom(),
@@ -108,13 +92,13 @@ abstract class DescopeBridgeHook(val events: Set<Event>) {
 
 /**
  * Disables two finger and double tap zoom gestures.
- *
- * This hook is always run automatically when the flow or widget is loaded, so there's
+ * 
+ * This hook is always run automatically when the flow is loaded, so there's
  * usually no need to use it in application code.
  *
  * @return The disable zoom hook
  */
-fun disableZoom(): DescopeBridgeHook {
+fun disableZoom(): DescopeFlowHook {
     return SetUpWebViewHook(setOf(Event.Started)) {
         it.settings.setSupportZoom(false)
     }
@@ -126,7 +110,7 @@ fun disableZoom(): DescopeBridgeHook {
  * @param setup The setup function will be called with an instance of `WebView`
  * @return The WebView setup hook
  */
-fun setUpWebView(setup: (WebView) -> Unit): DescopeBridgeHook {
+fun setUpWebView(setup: (WebView) -> Unit): DescopeFlowHook {
     return SetUpWebViewHook(setOf(Event.Started), setup)
 }
 
@@ -137,19 +121,19 @@ fun setUpWebView(setup: (WebView) -> Unit): DescopeBridgeHook {
  * @param code The code to run
  * @return The run JavaScript hook
  */
-fun runJavaScript(event: Event, code: String): DescopeBridgeHook {
+fun runJavaScript(event: Event, code: String): DescopeFlowHook {
     return RunJavaScriptHook(setOf(event), code)
 }
 
 /**
- * Add styles to the web page presented by the flow or widget
+ * Add styles to the web page presented by the flow
  *
  * @param event The event at which to run (Loaded by default)
  * @param selector The CSS selector
  * @param rules The CSS rules to apply
  * @return The add styles hook
  */
-fun addStyles(event: Event = Event.Loaded, selector: String, rules: List<String>): DescopeBridgeHook {
+fun addStyles(event: Event = Event.Loaded, selector: String, rules: List<String>): DescopeFlowHook {
     return AddStylesHook(
         setOf(event), css = """
 $selector {
@@ -172,32 +156,32 @@ fun setTransparentBody() = setBackgroundColor("body", 0x00000000)
  * @param color The color to apply as the `background-color`
  * @return The set background color hook
  */
-fun setBackgroundColor(selector: String, color: Long): DescopeBridgeHook {
+fun setBackgroundColor(selector: String, color: Long): DescopeFlowHook {
     return BackgroundColorHook(selector = selector, color = color)
 }
 
 // Internal
 
-private class SetUpWebViewHook(events: Set<Event>, private val fn: (WebView) -> Unit) : DescopeBridgeHook(events) {
-    override fun execute(event: Event, coordinator: Coordinator) {
+private class SetUpWebViewHook(events: Set<Event>, private val fn: (WebView) -> Unit) : DescopeFlowHook(events) {
+    override fun execute(event: Event, coordinator: DescopeFlowCoordinator) {
         fn(coordinator.webView)
     }
 }
 
-private class RunJavaScriptHook(events: Set<Event>, private val code: String) : DescopeBridgeHook(events) {
-    override fun execute(event: Event, coordinator: Coordinator) {
+private class RunJavaScriptHook(events: Set<Event>, private val code: String) : DescopeFlowHook(events) {
+    override fun execute(event: Event, coordinator: DescopeFlowCoordinator) {
         coordinator.runJavaScript(code)
     }
 }
 
-private class AddStylesHook(events: Set<Event>, private val css: String) : DescopeBridgeHook(events) {
-    override fun execute(event: Event, coordinator: Coordinator) {
+private class AddStylesHook(events: Set<Event>, private val css: String) : DescopeFlowHook(events) {
+    override fun execute(event: Event, coordinator: DescopeFlowCoordinator) {
         coordinator.addStyles(css)
     }
 }
 
-private class BackgroundColorHook(events: Set<Event> = setOf(Event.Loaded), private val selector: String, private val color: Long) : DescopeBridgeHook(events) {
-    override fun execute(event: Event, coordinator: Coordinator) {
+private class BackgroundColorHook(events: Set<Event> = setOf(Event.Loaded), private val selector: String, private val color: Long) : DescopeFlowHook(events) {
+    override fun execute(event: Event, coordinator: DescopeFlowCoordinator) {
         coordinator.addStyles("$selector { background-color: $colorStringValue; }")
     }
