+++ descopesdk/src/main/java/com/descope/internal/others/Errors.kt
@@ -1,7 +1,6 @@
 package com.descope.internal.others
 
 import com.descope.types.DescopeException
-import org.json.JSONObject
 
 internal fun DescopeException.with(desc: String? = null, message: String? = null, cause: Throwable? = null) = DescopeException(
     code = code,
@@ -9,15 +8,3 @@ internal fun DescopeException.with(desc: String? = null, message: String? = null
     message = message ?: this.message,
     cause = cause ?: this.cause,
 )
-
-internal fun parseServerError(response: String): DescopeException? {
-    try {
-        val map = JSONObject(response).toMap()
-        val code = map["errorCode"] as? String ?: return null
-        val desc = map["errorDescription"] as? String ?: "Descope server error"
-        val message = map["errorMessage"] as? String
-        return DescopeException(code = code, desc = desc, message = message)
-    } catch (_: Exception) {
-        return null
-    }
-}
